apps/fastapi/requirements.txt
black
flake8
pandas
numpy
python-socketio
fastapi
pydantic
uvicorn
sqlalchemy
uvicorn
python-dotenv
dependency_injector
